add listdatareader
